requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.36
# psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==23.0.0
Werkzeug==3.0.1
